requirements.txt
# Flask & Extensions
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5

# Database
SQLAlchemy>=2.0.23

# Authentication & Security
Werkzeug>=3.0.1
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.3.0

# AI Integration
requests>=2.31.0
httpx>=0.27.0

# File Handling
python-magic>=0.4.27

# Production Server
gunicorn>=21.2.0

# Development
pytest>=8.0.0
pytest-flask>=1.3.0
black>=24.0.0
ruff>=0.1.0
pre-commit>=3.6.0
